fix: header menu layout is getting closer but not 100% yet
