+++ blocks/breadcrumb/breadcrumb.css
@@ -11,6 +11,7 @@ main .breadcrumb-container {
 }
 
 .breadcrumb > nav {
+  font-family: 'Helvetica Neue LT W05_65 Medium', Helvetica, Arial, sans-serif;
   color: var(--color-deepblue-10);
   padding: 2rem;
   line-height: 1.75rem;
@@ -27,7 +28,7 @@ main .breadcrumb-container {
 }
 
 .breadcrumb .breadcrumb-separator {
-  margin-inline: 1.5ch;
+  margin-inline: 1ch;
 }
 
 .breadcrumb .breadcrumb-separator::after {
